requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==2.3.1
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
